Enable sample/simple for aarch64 (#616)
Fixed a typo in aot_reloc_aarch64.c.
Remove -m32 option in host tool CMakeLists.txt and change data type of structure fields to make host tool work in 64-bit.

+++ test-tools/host-tool/src/transport.h
@@ -13,7 +13,7 @@ extern "C" {
 /* IMRT link message between host and WAMR */
 typedef struct {
     unsigned short message_type;
-    unsigned long payload_size;
+    unsigned int payload_size;
     char *payload;
 } imrt_link_message_t;
